pdf-worker/requirements.txt
redis==5.0.8
pydantic-settings==2.4.0
weasyprint==61.2
pydyf==0.8.0
jinja2==3.1.4
google-api-python-client==2.136.0
google-auth==2.32.0
google-auth-oauthlib==1.2.1
